requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyYAML==6.0
uritemplate==4.1.1
django-parler==2.2.1
django-parler-rest==2.1
drf-yasg==1.20.0
django-rest-swagger==2.2.0